Consolidated InterpolateQPVelocity_* and InterpolateQPAcceleration_* into InterpolateQPVector
No performance change, but 1/4 the code.  Subviews are used here, but are hidden in the
InterpolateToQuadraturePoints kernel, so the top level API is still pretty clean.
