+++ src/restruct_poc/beams/interpolate_to_quadrature_points.hpp
@@ -4,8 +4,7 @@
 
 #include "beams.hpp"
 #include "interpolate_QP_state.hpp"
-#include "interpolate_QP_velocity.hpp"
-#include "interpolate_QP_acceleration.hpp"
+#include "interpolate_QP_vector.hpp"
 
 #include "src/restruct_poc/types.hpp"
 
@@ -39,10 +38,11 @@ struct InterpolateToQuadraturePoints {
         Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_uprime{first_qp, first_node, num_nodes, shape_deriv, qp_jacobian, node_u, qp_uprime});
         Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_r{first_qp, first_node, num_nodes, shape_interp, node_u, qp_r});
         Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPState_rprime{first_qp, first_node, num_nodes, shape_deriv, qp_jacobian, node_u, qp_rprime});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVelocity_Translational{first_qp, first_node, num_nodes, shape_interp, node_u_dot, qp_u_dot});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVelocity_Angular{first_qp, first_node, num_nodes, shape_interp, node_u_dot, qp_omega});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPAcceleration_Translational{first_qp, first_node, num_nodes, shape_interp, node_u_ddot, qp_u_ddot});
-        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPAcceleration_Angular{first_qp, first_node, num_nodes, shape_interp, node_u_ddot, qp_omega_dot});
+
+        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVector{first_qp, first_node, num_nodes, shape_interp, Kokkos::subview(node_u_dot, Kokkos::ALL, Kokkos::pair(0, 3)), qp_u_dot});
+        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVector{first_qp, first_node, num_nodes, shape_interp, Kokkos::subview(node_u_dot, Kokkos::ALL, Kokkos::pair(3, 6)), qp_omega});
+        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVector{first_qp, first_node, num_nodes, shape_interp, Kokkos::subview(node_u_ddot, Kokkos::ALL, Kokkos::pair(0, 3)), qp_u_ddot});
+        Kokkos::parallel_for(Kokkos::TeamThreadRange(member, idx.num_qps), InterpolateQPVector{first_qp, first_node, num_nodes, shape_interp, Kokkos::subview(node_u_ddot, Kokkos::ALL, Kokkos::pair(3, 6)), qp_omega_dot});
     }
 };
