+++ tests/unit_tests/restruct_poc/CMakeLists.txt
@@ -10,7 +10,7 @@ target_sources(
     test_math.cpp
     test_utilities.cpp
     beams/test_interpolate_QP_state.cpp
-    beams/test_interpolate_QP_velocity.cpp
+    beams/test_interpolate_QP_vector.cpp
     solver/test_compute_number_of_non_zeros.cpp
     solver/test_copy_into_sparse_matrix.cpp
     solver/test_populate_sparse_indices.cpp
